requirements.txt
feedparser
google-generativeai
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
